Create servants.alternate.json
Finally understood why the alternate list is required -- to make Mash SR. Will be looking into just creating both entries in the main data file and making the code figure out which one to display based on the "Mash is SR" option. In the meantime readding the file to avoid issues.