Update index.js
Adjusted the URL shortening function to work with three staggered providers. If the first fails, the next one will take over, and so on until the last one. If the last one fails, a message will be shown to the user indicating that URL shortening is currently out of service.